hotfix fix coder. RMSNorm cannot have float16 input

+++ extra/models/llama.py
@@ -36,7 +36,7 @@ class RMSNorm:
     self.weight = Tensor.ones(dim)
 
   def __call__(self, x:Tensor):
-    # TODO: convert to float?
+    x = x.float()
     return (x * (x.pow(2).mean(-1, keepdim=True) + self.eps).rsqrt()) * self.weight
 
 class Attention:
